+++ src/application/Layout/NavBar/NavBar.css
@@ -2,7 +2,7 @@
     flex: 1;
     display: flex;
     justify-content: space-between;
-    padding: 10px;
+    padding: 10px 15px;
     background-color: #ffcb05;
     align-items: center;
     border-bottom: 2px solid black;
