+++ src/components/ErrorBoundary/ErrorBoundary.css
@@ -9,7 +9,7 @@
 .error-image {
     width: 300px;
     height: 300px;
-    background-image: url('../../../assets/image/pikachu.jpg');
+    background-image: url('../../assets/image/pikachu-error.jpg');
     background-size: 100%;
     background-repeat: no-repeat;
 }
