+++ src/components/ErrorBoundary/type.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react';
+
+export interface ErrorBoundaryState {
+    hasError: boolean;
+}
+
+export interface ErrorBoundaryProps {
+    children: ReactNode;
+}
