+++ src/components/ErrorButton/type.tsx
@@ -1,0 +1,3 @@
+export interface ErrorButtonState {
+    hasError: boolean;
+}
