+++ src/pages/AboutPage/AboutPage.css
@@ -18,7 +18,7 @@
 .about-img {
     width: 300px;
     height: 300px;
-    background-image: url('../../../assets/image/pikachy.png');
+    background-image: url('../../assets/image/pikachy.png');
     background-size: 100%;
     background-repeat: no-repeat;
 }
