+++ src/pages/CardsPage/components/CardsField/type.tsx
@@ -1,0 +1,10 @@
+import { PokemonCard } from '../../../../types/types';
+
+export interface CardsFieldProps {
+    isDataLoaded: boolean;
+    cardsData: PokemonCard;
+}
+export interface CardsFieldState {
+    isDataLoaded: boolean;
+    cardsData: PokemonCard;
+}
