+++ src/pages/CardsPage/components/SearchForm/type.tsx
@@ -1,0 +1,7 @@
+export interface SearchFormProps {
+    onValueChange: (value: string) => void;
+}
+
+export interface SearchFormState {
+    inputValue: string;
+}
