+++ src/pages/CardsPage/type.tsx
@@ -1,0 +1,7 @@
+import { PokemonCard } from '../../types/types';
+
+export interface CardsPageState {
+    inputValue: string;
+    isDataLoaded: boolean;
+    cardsData: PokemonCard;
+}
