+++ src/types/types.tsx
@@ -1,13 +1,21 @@
-import { ReactNode } from 'react';
-
-export interface SearchFormProps {
-    onValueChange: (value: string) => void;
+export interface PokemonCard {
+    data: Array<Pokemon>;
+    page: number;
+    pageSize: number;
+    count: number;
+    totalCount: number;
 }
-
-export interface SearchFormState {
-    inputValue: string;
+interface Pokemon {
+    id: string;
+    name: string;
+    abilities: Array<Ability>;
+    attacks: Array<Attacks>;
+    hp: number;
+    images: {
+        small: string;
+        large: string;
+    };
 }
-
 interface Ability {
     name: string;
     text: string;
@@ -20,51 +28,4 @@ interface Attacks {
     convertedEnergyCost: number;
     damage: string;
     text: string;
-}
-
-interface Pokemon {
-    id: string;
-    name: string;
-    abilities: Array<Ability>;
-    attacks: Array<Attacks>;
-    hp: number;
-    images: {
-        small: string;
-        large: string;
-    };
-}
-
-export interface PokemonCard {
-    data: Array<Pokemon>;
-    page: number;
-    pageSize: number;
-    count: number;
-    totalCount: number;
-}
-
-export interface CardsFieldProps {
-    isDataLoaded: boolean;
-    cardsData: PokemonCard;
-}
-export interface CardsFieldState {
-    isDataLoaded: boolean;
-    cardsData: PokemonCard;
-}
-
-export interface CardsPageState {
-    inputValue: string;
-    isDataLoaded: boolean;
-    cardsData: PokemonCard;
-}
-
-export interface ErrorBoundaryState {
-    hasError: boolean;
-}
-
-export interface ErrorButtonState {
-    hasError: boolean;
-}
-
-export interface ErrorBoundaryProps {
-    children: ReactNode;
 }
